Repeat initial state messages when the HPB is used
When the HPB is used it is possible to know when the local participant
has established a connection with a remote participant, but it is not
possible to know when a remote participant established a connection with
the local one. Currently the initial state (audio and video on or off)
is sent when the local participant establishes a connection with the
remote one, which causes that, in some cases, the message is received by
the remote participant before a connection was established with the
local one and thus ignored.

Even if there is code to try to locally detect the initial media state
of the remote participant it has shown to not be as reliable as
desirable, which causes a wrong media state to be sometimes shown for
remote participants until the participant changes it again (or another
user joins and the messages are sent again to everyone).

Even if it is not possible to know when a remote participant established
a connection with the local one it is safe to assume that it happened
in a "reasonable" amount of time since the local participant has
established a connection with the remote one.

Due to this now the initial media state is first sent when that happens,
and then it is sent again several times (with an increasing interval) in
the next 15 seconds. If a connection is established with another
participant in the meantime the cycle starts again.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>